indicators_live/indicators_live_main.py
# indicators_live_main.py
# Координатор live-расчёта технических индикаторов по mark price

# 0. Импорты
import asyncio
import asyncpg
import redis.asyncio as aioredis
import os
import json
import pandas as pd
from datetime import datetime

# 1. Переменные окружения
DATABASE_URL = os.getenv("DATABASE_URL")
REDIS_HOST = os.getenv("REDIS_HOST")
REDIS_PORT = int(os.getenv("REDIS_PORT", 6379))
REDIS_PASSWORD = os.getenv("REDIS_PASSWORD")

# 2. Основная точка входа
async def main():
    print("[INIT] indicators_live_main стартует", flush=True)
    await asyncio.sleep(1)

    pg_pool = None
    redis = None

    try:
        # 2.1 Подключение к PostgreSQL
        pg_pool = await asyncpg.create_pool(DATABASE_URL)
        print("[PG] Подключение к PostgreSQL установлено", flush=True)

        # 2.2 Подключение к Redis
        redis = aioredis.Redis(
            host=REDIS_HOST,
            port=REDIS_PORT,
            password=REDIS_PASSWORD,
            ssl=True
        )
        await redis.ping()
        print("[Redis] Подключение к Redis установлено", flush=True)

        # 2.3 Запуск главного цикла (заглушка)
        await run_live_loop(pg_pool, redis)

    except Exception as e:
        print(f"[ERROR] Инициализация завершилась с ошибкой: {e}", flush=True)

    finally:
        if pg_pool:
            await pg_pool.close()
        if redis:
            await redis.close()


# 3. Главный рабочий цикл (заглушка)
async def run_live_loop(pg_pool, redis):
    print("[LOOP] Запуск основного цикла обработки индикаторов", flush=True)
    while True:
        await asyncio.sleep(2)
        # TODO: здесь будет логика обновления mark price, расчёт индикаторов и публикация в Redis
        print("[LOOP] Псевдо-расчёт завершён", flush=True)


# 4. Запуск
if __name__ == "__main__":
    asyncio.run(main())
