indicators_live/requirements.txt
# Асинхронные библиотеки
asyncpg==0.29.0
aioredis==2.0.1
aiohttp==3.9.3
websockets==12.0
python-dotenv==1.0.1

# Для расчётов и дат
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.8.2
pytz==2024.1
pendulum==3.0.0

# (Опционально, если будешь использовать uvloop для ускорения)
uvloop==0.19.0